# Make uptime generation optional