I'd like my benchmark tool to create the exact number of packets it says it will; the extra uptime thing is not necesarry for what I'm doing, so I added an option to turn it off.

# Testing 
## Without `-uptimeenable` we send exactly 1 message per batch
```
[csibbitt@laptop telemetry-bench (csibbitt-make-uptime-optional) ]$ ./telemetry-bench -hostprefix test054- -hosts 1 -plugins 1 -instances 1 -verbose -timepermesgs 1 -send 2 -interval 1 amqp://localhost:5672/collectd/telemetry/
Send 1 metrics every 1 second(s)Total sent (0)0, total 0, 0 ack'd
Generated 1 metrics in 310.856µs
(0): Sent 1 metrics in 762.978582ms, ( 762978.582 uS per metric )
Total sent (0)1, total 1, 1 ack'd
Generated 1 metrics in 215.545µs
(0): Sent 1 metrics in 77.173µs, ( 77.173 uS per metric )
done...
``` 
## With `-uptimeenable` we send 2 messages per batch
```
[csibbitt@laptop telemetry-bench (csibbitt-make-uptime-optional) ]$ ./telemetry-bench -hostprefix test055- -hosts 1 -plugins 1 -instances 1 -verbose -timepermesgs 1 -send 2 -interval 1 -uptimeenable amqp://local
host:5672/collectd/telemetry/
Send 1 metrics every 1 second(s)Total sent (0)0, total 0, 0 ack'd
Generated 2 metrics in 666.808µs
(0): Sent 1 metrics in 781.763239ms, ( 781763.239 uS per metric )
(0): Sent 1 metrics in 388.377µs, ( 388.377 uS per metric )
Total sent (0)2, total 2, 2 ack'd
Generated 2 metrics in 290.356µs
(0): Sent 1 metrics in 78.575µs, ( 78.575 uS per metric )
(0): Sent 1 metrics in 263.718µs, ( 263.718 uS per metric )
done...
```
I checked in prometheus, and in the second case I see the uptime metric, whereas in the first case I don't.