@csibbitt 2019-09-13
This will conflict with my other PR #8 . It will be easiest to merge that one first I think.